vdisp/deltacast: set correct Amercian system
Althought there are in deltacast_frame_modes the modes for eg. 1080p29
and 1080p30 listed separately, both point to the same video standard
mode VHD_VIDEOSTD_S274M_1080p_30Hz. This is perhaps taken orignally from
some SDK samples, anyways. The clock system seem to be required to set
separate from the video standard. This was hard-coded in init to 1/1.